消息改为二进制传输

+++ src/main/java/com/quwan/im/model/MessageType.java
@@ -57,7 +57,7 @@ public enum MessageType {
      * 获取消息类型编码
      * @return 编码值
      */
-    public int getCode() {
+    public byte getCode() {
         return code;
     }
